Feat/ATM-87 USDT on XTZ (#54)
* Add USDT_XTZ Fa2 token

* Remove unused image resources, add UsdtXtzCurrencyViewModel

* Fix Wert currencies icons

* Add Fa2WalletView & ViewModel

* Add Fa2Config usage

* Unify Fa12Account and Fa2Account

* Fix TezosToken view models

* Fix Fee for Fa2SendViewModel, fix TezosTokensSendViewModel

* Fix all displayed names

* Increase app version to 1.2.15

* Update client core library

* Fix TokenId

* Fix Tezos tokens codes

+++ ViewModels/WalletViewModels/TezosTokensWalletViewModel.cs
@@ -432,7 +432,7 @@ namespace Atomex.Client.Desktop.ViewModels.WalletViewModels
             var sendViewModel = new TezosTokensSendViewModel(
                 app: _app,
                 tokenContract: tokenViewModel.TokenBalance.Contract,
-                tokenId: tokenViewModel.TokenBalance.TokenId,
+                tokenId: (int)tokenViewModel.TokenBalance.TokenId,
                 tokenType: TokenContract.Contract.GetContractType(),
                 // getTokenPreview: GetTokenPreview,
                 tokenPreview: null,
